backend/managers: Correcting datatype of funct options
While reading the code below, it seems like we are passing strings for both options here.
`rename` enforces the download filename, which is no Pathlib kind of variable, but pure string.
And `checksum` is an md5-checksum, which is a string due to usage of hexdigest.

Therefore correcting the types here, as intended to be used in the function.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com>