backend/managers: Hardening checksum comparision
Changed the fallback from False to None in the hasher since it makes more sense to me that some hashing gave no results then something wrong(?).
We are asking the function for a checksum and not for the result of our (later) comparison.
While changing this, I also added a check in the download code to check for None, so we don't run into a None vs. string comparison.

This should make the checksum checking more robust.

NOTE: I made no check, whether the change in the hashing function influences the logic somewhere else!

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com>