backend/managers: Move file_path outside if clause
I get scared everytime I see code like this.
If someone overlooks the first-time-declaration inside the if-clause, then we might run into errors later because of undeclared variables.
It is not a problem here, since we have not too much code, but I see it as a common practice to avoid this type of mistake systematically.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com>