Remember last used view in topmost `AdwViewStack`
It makes sense to start the application in the main view where it was
closed. For example, it is plausible that most users set up their
bottles once and then want to land directly in the library view after
launch.

This behavior is also known from other programs such as web browsers or
GNOME Calendar, and it seems that Bottles once handled it this way in
the past. Therefore the gsettings key `startup-view` has now been
(re)activated and represents the last selected view in the main window.