backend: workaround #2062
This is just a workaround. I'm not able to find what is causing this TypeError, I've tested with some different setups and I've never been able to reproduce it, so I'm just providing a workaround for not breaking the app. Thiw workaround removes the path from the list of found dlls as it seems to empty in some cases.