Add command-line arguments
This improvement adds command line arguments
and help pages to the build script, extending
its functionality.

Arguments can now be passed through to the
flatpak, to gain access to bottles-cli at
runtime.

Additionally, the script now has help pages:

The help page can be accessed by running
`./build.sh --help` or `./build.sh -h`.

From there help on actions and options can be
viewed by running `./build.sh <action> --help.

NOTE: This script is almost generic enough
to be used for any flatpak builder project, but
it is still tailored to bottles as the
package name is hardcoded in the script.

It does not fully support all arguments of
flatpak itself (yet), but it is a good start.