Complex -f in ``list``, and additions
This commit adds complex filters to the ``-f|--filter``
parameter of the list action in bottles-cli.

The expanded functionality allows you to filter
by multiple fields and values, and to use
glob patterns to match.

Additionally, added the missing ability to
``list dependencies``, and migrated the programs
action over to the list action, where the
complex filter ``bottle:`` can be used to
target listing programs only found in a
single bottle (or multiple with glob match)