chore: Clean up frontend
This cleans up a considerable chunk of the codebase by removing the dedicated sandbox. As it stands, we only support sandboxed formats, so we might as well delete the unmaintained one.

This also removes a lot of unnecessary checks in regards to packaging. Since the Flatpak includes everything we need, we might as well just not have these checks anymore.

Some of the checks early return.