new-bottle-dialog: Split pages
The dialog is split into 3 pages as opposed to 2. This makes
transitioning between pages smoother, as it switches pages
with a crossfade transition as opposed to replacing widgets
and labels.

The scrolled window displaying the output is moved to the
completed page if there is an error, to keep the same functionality
as the previous implementation.

The second and last pages can also be dragged from anywhere,
except the scrolled output.