new-bottle-dialog: Rework widget sizing
This makes the dialog a bit narrower and increases the content
height. The creating page has gotten a small rework to accommodate
the height. The output label is bigger and fits better with
different screen sizes.

The status page was replaced with a GtkScrolledWindow, as
implementing it ourselves gave us the control desired to better
control the height of the output label without hardcoding a value.