envvars-dialog: Center text entry vertically
Traverse AdwEntryRow's widget tree to remove visibility of the title labels and change alignment of the text entry to center.

This implementation is tightly coupled with the internal representation of AdwEntryRow. Checks were added to prevent crashing if that representation changes, but in such case the layout may get askew.