Use `add_pattern` for new dialog's FileFilters
This addresses an issue originally fixed in #3221 where the GTK function
`add_media_type` acted inconsistently across distributions. This was
reintroduced in later releases after the UI was updated for creating
bottles.

As was before, the `add_media_type` function would be the ideal one to
use, so this is in actuality a temporary fix.