wrap runners/cache/dll preferences pages in AdwPreferencesGroup
These pages became blank when run with libadwaita 1.8.2/1.7.9 due to
https://gitlab.gnome.org/GNOME/libadwaita/-/commit/567552ed7876aed3e6e9a102824258baf4ba048f
which enforces the API contract that only AdwPreferencesGroup may be
added as a child to AdwPreferencesPage, when loading a designer file.

As a minimal fix, wrap the Stack elements in the preferences pages in
anonymous AdwPreferencesGroup elements so the pages appear again. The
functionality of the pages does not appear to be affected.

Fixes #4229.